requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
openai==1.14.2
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4
